Feature: add properties filter in ListServiceDetail API
