+++ datasource/service_test.go
@@ -571,152 +571,6 @@ func TestService_Delete(t *testing.T) {
 	})
 }
 
-func TestService_Info(t *testing.T) {
-	var (
-		serviceID  string
-		instanceID string
-	)
-	ctx := getContext()
-	defer datasource.GetMetadataManager().UnregisterService(ctx, &pb.DeleteServiceRequest{ServiceId: serviceID, Force: true})
-
-	t.Run("register a tested service & instance, should be passed", func(t *testing.T) {
-		resp, err := datasource.GetMetadataManager().RegisterService(ctx, &pb.CreateServiceRequest{
-			Service: &pb.MicroService{
-				ServiceName: "TestServic1",
-				AppId:       "default",
-				Version:     "1.0.0",
-				Level:       "FRONT",
-				Status:      "UP",
-			},
-		})
-		assert.NoError(t, err)
-		serviceID = resp.ServiceId
-
-		instance := &pb.MicroServiceInstance{
-			ServiceId: serviceID,
-			Endpoints: []string{
-				"rest://[::]:301000/",
-			},
-			HostName: "delete-host-ms",
-			Status:   pb.MSI_UP,
-		}
-		respInstance, err := datasource.GetMetadataManager().RegisterInstance(ctx, &pb.RegisterInstanceRequest{
-			Instance: instance,
-		})
-		assert.NoError(t, err)
-		instanceID = respInstance.InstanceId
-	})
-
-	t.Run("get tested service info, should be passed", func(t *testing.T) {
-		resp, err := datasource.GetMetadataManager().ListServiceDetail(ctx, &pb.GetServicesInfoRequest{
-			Options:     []string{"all"},
-			AppId:       "default",
-			ServiceName: "TestServic1",
-		})
-		assert.NoError(t, err)
-		assert.NotEqual(t, 0, len(resp.AllServicesDetail))
-		assert.NotEqual(t, 0, resp.Statistics.Services.Count)
-		assert.NotEqual(t, 0, resp.Statistics.Instances.Count)
-		assert.NotEqual(t, 0, resp.Statistics.Apps.Count)
-	})
-
-	t.Run("unregister tested service & instance, should be passed", func(t *testing.T) {
-		err := datasource.GetMetadataManager().UnregisterInstance(ctx, &pb.UnregisterInstanceRequest{
-			ServiceId:  serviceID,
-			InstanceId: instanceID,
-		})
-		assert.NoError(t, err)
-
-		err = datasource.GetMetadataManager().UnregisterService(ctx, &pb.DeleteServiceRequest{
-			ServiceId: serviceID,
-			Force:     true,
-		})
-		assert.NoError(t, err)
-	})
-
-	t.Run("get all services", func(t *testing.T) {
-		log.Info("should be passed")
-		resp, err := datasource.GetMetadataManager().ListServiceDetail(ctx, &pb.GetServicesInfoRequest{
-			Options: []string{"all"},
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, pb.ResponseSuccess, resp.Response.GetCode())
-
-		resp, err = datasource.GetMetadataManager().ListServiceDetail(ctx, &pb.GetServicesInfoRequest{
-			Options: []string{""},
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, pb.ResponseSuccess, resp.Response.GetCode())
-
-		resp, err = datasource.GetMetadataManager().ListServiceDetail(ctx, &pb.GetServicesInfoRequest{
-			Options: []string{"tags", "rules", "instances", "schemas", "statistics"},
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, pb.ResponseSuccess, resp.Response.GetCode())
-
-		resp, err = datasource.GetMetadataManager().ListServiceDetail(ctx, &pb.GetServicesInfoRequest{
-			Options: []string{"statistics"},
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, pb.ResponseSuccess, resp.Response.GetCode())
-
-		resp, err = datasource.GetMetadataManager().ListServiceDetail(ctx, &pb.GetServicesInfoRequest{
-			Options:   []string{"instances"},
-			CountOnly: true,
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, pb.ResponseSuccess, resp.Response.GetCode())
-	})
-}
-
-func TestService_Detail(t *testing.T) {
-	var (
-		serviceId string
-	)
-	ctx := getContext()
-	defer datasource.GetMetadataManager().UnregisterService(ctx, &pb.DeleteServiceRequest{ServiceId: serviceId, Force: true})
-
-	t.Run("execute 'get detail' operation", func(t *testing.T) {
-		log.Info("should be passed")
-		resp, err := datasource.GetMetadataManager().RegisterService(ctx, &pb.CreateServiceRequest{
-			Service: &pb.MicroService{
-				AppId:       "govern_service_group",
-				ServiceName: "govern_service_name",
-				Version:     "3.0.0",
-				Level:       "FRONT",
-				Status:      pb.MS_UP,
-			},
-		})
-		assert.NoError(t, err)
-		serviceId = resp.ServiceId
-
-		_, err = datasource.GetMetadataManager().ModifySchema(ctx, &pb.ModifySchemaRequest{
-			ServiceId: serviceId,
-			SchemaId:  "schemaId",
-			Schema:    "detail",
-		})
-		assert.NoError(t, err)
-
-		_, err = datasource.GetMetadataManager().RegisterInstance(ctx, &pb.RegisterInstanceRequest{
-			Instance: &pb.MicroServiceInstance{
-				ServiceId: serviceId,
-				Endpoints: []string{
-					"govern:127.0.0.1:8080",
-				},
-				HostName: "UT-HOST",
-				Status:   pb.MSI_UP,
-			},
-		})
-		assert.NoError(t, err)
-
-		err = datasource.GetMetadataManager().UnregisterService(ctx, &pb.DeleteServiceRequest{
-			ServiceId: serviceId,
-			Force:     true,
-		})
-		assert.NoError(t, err)
-	})
-}
-
 func TestApplication_Get(t *testing.T) {
 	t.Run("execute 'get apps' operation", func(t *testing.T) {
 		log.Info("when request is valid, should be passed")
